+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+* Create `BakeAppendixFeatureTitles` direction to allow appendices feature sections titles be dynamic
 * Add support for baking injected exercises/solutions in appendices
 ## [v1.7.0] - 2022-05-06