+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Change `references title` from `Endnotes` to `References` for `marketing`
 * Create `BakeAppendixFeatureTitles` direction to allow appendices feature sections titles be dynamic
 * Add `BakeFirstElements` direction to `college-physics`, `college-physics-2e` recipes
 * Add support for baking injected exercises/solutions in appendices