+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * Add support for baking learning objectives in appendices (`BakeLearningObjectives`) for `marketing`
+* Fix bad encoding in AP physics locales
+* Add `BakeFirstElements` to examples in `dev-math`
 * Change `references title` from `Endnotes` to `References` for `marketing`
 * Create `BakeAppendixFeatureTitles` direction to allow appendices feature sections titles be dynamic
 * Add `BakeFirstElements` direction to `college-physics`, `college-physics-2e` recipes